Modularization of get route, creation of id, name filters
