+++ api/controllers/news/get/getNewsById.js
@@ -1,0 +1,5 @@
+const news = require("../../../models/news");
+
+const getNewsByID = (req, res) => {};
+
+module.exports = { getNewsByID };
