+++ api/controllers/news/get/getNewsByName.js
@@ -1,0 +1,5 @@
+const news = require("../../../models/news");
+
+const getNewsByName = (req, res) => {};
+
+module.exports = { getNewsByName };
